Modelo + Enums + Algunos Endpoints

+++ Backend/src/main/java/com/example/cartasIntercambio/dto/OfertaDto.java
@@ -1,6 +1,8 @@
 package com.example.cartasIntercambio.dto;
 
+import com.example.cartasIntercambio.model.Mercado.EstadoOferta;
 import com.example.cartasIntercambio.model.Producto_Carta.Carta;
+import com.example.cartasIntercambio.model.Producto_Carta.EstadoCarta;
 import com.example.cartasIntercambio.model.Usuario.Usuario;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -8,6 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -19,17 +22,16 @@ public class OfertaDto {
     
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private Date fecha;
-    //private Publicacion publicacion;
 
     private Long idPublicacion;
-    private Double monto;
+    private BigDecimal monto;
     private List<Carta> cartasOfrecidas;
     private Usuario ofertante;
+    //private Long idOfertante;
     private String estado;
 
-  public OfertaDto(Date fecha, /*Publicacion publicacion*/ Long idPublicacion, Double monto, List<Carta> cartasOfrecidas, Usuario ofertante, String estado) {
+  public OfertaDto(Date fecha, Long idPublicacion, BigDecimal monto, List<Carta> cartasOfrecidas, Usuario ofertante, String estado) {
     this.fecha = fecha;
-    //this.publicacion = publicacion;
     this.idPublicacion = idPublicacion;
     this.monto = monto;
     this.cartasOfrecidas = cartasOfrecidas;
